refactor(core)!: the loop's events are San's, not a translation of them
core.Event was a bus that predates the SDK: a string Type, a Source nobody
read, and a Data any that every consumer opened with an ok it then dropped.
translate() spent fifty lines turning the loop's twelve typed events into it —
a second vocabulary for the same twelve facts, which is a thing that can be
wrong, and was: a refusal mapped onto StopError until two commits ago.

The outbox carries the loop's events now, as they arrive.

	switch e := ev.(type) {
	case sdkagent.MessageEnd:      // e.Response, not ev.Response() and an ok
	case sdkagent.MessageStart:    // e.Attempt, not an event San derived
	case sdkagent.CompactionEnd:   // e.Err, not a copy of it
	case core.TurnEnded:           // San's, because a turn is the mailbox's
	}

Seventeen event types become seven, and four of those the SDK has no concept
of because they are about the mailbox and the registries rather than one
exchange: AgentStarted/AgentStopped bracket Run, MessageReceived is a message
arriving in the inbox, TurnEnded closes however many exchanges a turn drained
into. Compacted stays because the summary text and the ID a transcript replays
from are the application's, and SystemChange/ToolsChange because a registry
mutation is not something the loop can see.

Three of San's went away entirely rather than being ported:

  StreamReset       derived from MessageStart.Attempt > 1. A consumer that has
                    the attempt number does not need San to have counted for it.
  CompactStart/End  the SDK's span, forwarded verbatim.
  PreInfer          San sent digests instead of the inference, to keep an
                    observer from copying the conversation every turn. Only the
                    transcript ever wanted them, so the digesting moved there,
                    with digest.go, and the UI reads the inference it was
                    already being handed.

What is left in exchange.go is folding, not translating: the Result is built
out of the same stream a consumer sees, so what an observer saw and what
ThinkAct returns cannot come to disagree about one turn.
