+++ docs/concepts/rendering.zh.md
@@ -187,7 +187,7 @@ InlinedResults.IsResultInlined(idx)      // 这条 result 是否会被
 ### Step 1 — PreInfer：开一个空的 assistant 桩位
 
 ```
-event:           core.PreInfer
+event:           sdkagent.MessageStart
 applyPreInfer:   m.Stream.Active = true
                  m.Append({Role: assistant, Content: ""})
                  启动 spinner
@@ -220,7 +220,7 @@ View → renderNormalView
 ### Step 2 — OnChunk（文字）：消息变长
 
 ```
-event:           core.OnChunk{Text: "我用 ls 列一下。", Done: false}
+event:           sdkagent.MessageUpdate{Delta: {Block: {Type: text, Text: "我用 ls 列一下。"}}}
 applyChunk:      m.AppendToLast(text, "")
 
 conv.Messages:   [user, assistant{Content:"我用 ls 列一下。"}]
@@ -234,7 +234,7 @@ Stream.Active:   仍然 true（Done=false）
 ### Step 3 — PostInfer：工具调用挂到 assistant 消息上
 
 ```
-event:           core.PostInfer{Response: {ToolCalls: [{ID:"tc-1", Name:"Bash", Input:{cmd:"ls"}}]}}
+event:           sdkagent.MessageEnd{Response: {Content: [ToolCall{ID:"tc-1", Name:"Bash", Input:{cmd:"ls"}}]}}
 applyPostInfer:  rt.OnInference(resp)
                  m.SetLastToolCalls(resp.ToolCalls)
                  m.Tool.Track(resp.ToolCalls)
@@ -270,7 +270,7 @@ RenderToolCalls(ToolCallsParams{
 ### Step 4 — PostTool：结果到了，配对 inline
 
 ```
-event:           core.PostTool{Result: {ToolCallID:"tc-1", Content:"file1\nfile2"}}
+event:           sdkagent.ToolEnd{ID:"tc-1", Result: {Content:"file1\nfile2"}}
 m.ProcessToolResult(tr):
   applyToolSideEffects(...)
   firePostToolHook(...)
@@ -313,7 +313,7 @@ RenderMessageRange(ctx, 1, 3, includeSpinner=true):
 ### Step 5 — OnChunk(Done)：把整块送进 scrollback
 
 ```
-event:           core.OnChunk{Done: true, Response: {...}}
+event:           sdkagent.MessageEnd{Response: {...}}
 applyChunk:      m.AppendToLast(...)       （可能还有最后一段文字 chunk）
                  if chunk.Done && 没有未完成的工具调用:
                      m.Stream.Active = false
