+++ internal/app/model_token_test.go
@@ -114,7 +114,7 @@ func TestAgentMessageEchoIsIgnored(t *testing.T) {
 		services:  services{Tracker: todo.NewStore(), Agent: &agent.Session{}},
 	}
 
-	echo := core.MessageEvent("agent-1", core.UserMessage("anything", nil))
+	echo := core.MessageReceived{Message: core.UserMessage("anything", nil)}
 	cmd, handled := conv.Update(m, &m.conv, conv.AgentOutboxMsg{Event: echo})
 
 	if !handled {
