+++ internal/core/run.go
@@ -3,6 +3,7 @@ package core
 import (
 	"context"
 	"errors"
+	"fmt"
 	"log"
 	"sync"
 	"sync/atomic"
@@ -14,15 +15,16 @@ import (
 	glog "github.com/genai-io/san/internal/log"
 )
 
-// agent drives sdk-go's loop and keeps San's shape around it.
+// agent is the mailbox that waits for work and the exchange it hands each
+// batch to. One file because they are one struct.
 //
-// What stays here is the mailbox: Run blocks on an inbox, drains what
-// accumulated, and hands the batch to one exchange — a shape the SDK
-// deliberately does not have, because how messages are batched into exchanges
-// is the application's to decide. What is no longer here is everything inside
-// one exchange: the inference and its retries, the tool batch and its
-// parallelism, compaction. That is sdkagent.Agent, and this file is the
-// translation between its events and San's.
+// The mailbox is the shape the SDK deliberately does not have: how messages
+// are batched into exchanges is the application's to decide. Inside an
+// exchange — the inference and its retries, the tool batch and its
+// parallelism, compaction — is sdkagent.Agent's, and ThinkAct drives it.
+//
+// Two things an exchange asks for live elsewhere: what the model may call is
+// toolset.go, and shortening the conversation is compact.go.
 type agent struct {
 	id          string
 	system      System
@@ -46,10 +48,9 @@ type agent struct {
 
 	closed atomic.Bool // guards outbox writes after close
 
-	// turn captures the in-flight ThinkAct so InterruptCurrentTurn can
-	// cancel it and the caller can wait for it to actually return. Stored
-	// as a pointer so swap-with-nil acts as an atomic claim — concurrent
-	// InterruptCurrentTurn calls become no-ops.
+	// turn is the in-flight ThinkAct, so InterruptCurrentTurn can cancel it and
+	// wait for it to unwind. A pointer, so swap-with-nil is an atomic claim and
+	// concurrent interrupts become no-ops.
 	turn atomic.Pointer[turnHandle]
 
 	// interruptPending latches an interrupt that arrived while Run was
@@ -88,13 +89,13 @@ func (a *agent) Append(ctx context.Context, msg Message) {
 }
 
 func (a *agent) Run(ctx context.Context) error {
-	a.emit(ctx, StartEvent(a.id))
+	a.emit(ctx, AgentStarted{})
 
 	var runErr error
 	defer func() {
 		// StopEvent must be delivered even on context cancellation,
 		// so use emitFinal which bypasses ctx.Done().
-		a.emitFinal(StopEvent(a.id, runErr))
+		a.emitFinal(AgentStopped{Err: runErr})
 		a.closed.Store(true)
 
 		if a.outbox != nil {
@@ -127,11 +128,11 @@ func (a *agent) Run(ctx context.Context) error {
 			}
 
 			// A failed turn (StopError) is an agent stop, not a turn boundary:
-			// emitting TurnEvent would fire OnTurnEnd on top of OnAgentStop.
+			// emitting TurnEnded would fire OnTurnEnd on top of OnAgentStop.
 			// Cancellation still emits (OnTurnEnd guards StopCanceled).
 			if result != nil && result.StopReason != StopError {
-				glog.QueueLog("agent.Run: ThinkAct done, emitting TurnEvent")
-				a.emit(ctx, TurnEvent(a.id, *result))
+				glog.QueueLog("agent.Run: ThinkAct done, emitting TurnEnded")
+				a.emit(ctx, TurnEnded{Result: *result})
 			}
 			if err != nil {
 				glog.QueueLog("agent.Run: ThinkAct error: %v", err)
@@ -284,7 +285,7 @@ func (a *agent) ingest(ctx context.Context, in Inbound) bool {
 	if in.Signal != "" {
 		return false
 	}
-	a.emit(ctx, MessageEvent(a.id, in.Msg))
+	a.emit(ctx, MessageReceived{Message: in.Msg})
 
 	a.mu.Lock()
 	a.pending = append(a.pending, in.Msg)
@@ -350,7 +351,7 @@ func (a *agent) emitFinal(event Event) {
 	select {
 	case a.outbox <- event:
 	case <-timer.C:
-		log.Printf("core/agent: failed to deliver %s event (outbox full for 5s)", event.Type)
+		log.Printf("core/agent: failed to deliver %T (outbox full for 5s)", event)
 	}
 }
 
@@ -372,4 +373,87 @@ func (a *agent) drainInbox(ctx context.Context) (int, error) {
 	default:
 		return 0, nil
 	}
+}
+
+// ThinkAct runs one exchange and reports what it produced. The Result is
+// folded out of the event stream rather than tracked alongside it, so what an
+// observer sees and what this returns cannot disagree about one turn.
+func (a *agent) ThinkAct(ctx context.Context) (*Result, error) {
+	// The prompt and the toolset are read afresh every exchange: both are
+	// registries the application mutates while the agent runs.
+	a.inner.SetSystem(a.system.Prompt())
+	a.inner.SetTools(a.offered()...)
+
+	out := &Result{}
+	var turnErr error
+
+	for event, err := range a.inner.Run(ctx, a.takePending()...) {
+		if err != nil {
+			// Outside a turn — ErrBusy today, which means two callers drove
+			// one conversation and the second must not silently do nothing.
+			return nil, err
+		}
+		a.fold(ctx, event, out)
+		if end, ok := event.(sdkagent.TurnEnd); ok {
+			turnErr = end.Err
+		}
+	}
+
+	out.Messages = a.Messages()
+	return out, turnErr
+}
+
+// fold puts the loop's event on the outbox, as it arrived, and folds it into
+// the outcome.
+func (a *agent) fold(ctx context.Context, event sdkagent.Event, out *Result) {
+	switch e := event.(type) {
+	case sdkagent.MessageEnd:
+		if e.Err == nil && e.Response != nil {
+			out.Steps++
+			out.Usage.Add(e.Response.Usage)
+		}
+
+	case sdkagent.ToolStart:
+		out.ToolUses++
+
+	case sdkagent.TurnEnd:
+		out.Content = e.Message.Text()
+		out.StopReason = e.StopReason
+		if e.Err != nil {
+			out.StopDetail = e.Err.Error()
+		}
+	}
+	a.emit(ctx, event)
+}
+
+// hooks is where San gets between the SDK's loop and the model: which client
+// answers this turn and what options ride with it, and — in compact.go, since
+// the answer is the same one /compact gives — when the conversation has grown
+// past what can be sent.
+func (a *agent) hooks() sdkagent.Hook {
+	return sdkagent.Hook{
+		PreInfer:     a.preInfer,
+		PreStep:      a.preStep,
+		OnInferError: a.onInferError,
+	}
+}
+
+// preInfer points the call at the client this turn's messages ask for, and
+// layers on whatever options the application wants with it.
+//
+// The client is asked for per call rather than held on the agent because one
+// vendor's headers depend on what the turn sends — an interactive endpoint
+// bills a user-initiated turn differently from an agent-initiated one.
+func (a *agent) preInfer(_ context.Context, inf *sdkagent.Inference) error {
+	if a.client != nil {
+		client, err := a.client(inf.Messages)
+		if err != nil {
+			return fmt.Errorf("reaching the model: %w", err)
+		}
+		inf.Client = client
+	}
+	if a.callOptions != nil {
+		inf.Options = append(inf.Options, a.callOptions()...)
+	}
+	return nil
 }
