+++ internal/session/recorder_lifecycle_test.go
@@ -36,9 +36,9 @@ func TestRecorder_WritesSessionStartedBeforeTelemetry(t *testing.T) {
 
 	// Fire a SystemChange right after — mirroring what SetObserver replay
 	// does inside core.NewAgent.
-	rec.OnAgentEvent(core.Event{Type: core.OnSystemChange, Data: core.SystemChange{
+	rec.OnAgentEvent(core.SystemChange{
 		Name: "identity", Slot: 0, Content: "You are X", Caller: "system:init",
-	}})
+	})
 
 	data, err := os.ReadFile(filepath.Join(dir, "transcripts", "sess.jsonl"))
 	if err != nil {
